Update: Laplace AI framework, knowledge integration, reasoning, robot agent, computer operation, learning modules
- Core: laplace_unified (h/c), loss.h, laplace_ai_framework, laplace_enhanced, laplace_integration, cfc_cell, cfc_enhanced, dynamics, lnn, optimizer, selflnn, tensor, CMakeLists.txt, msvc_stubs.c
- Knowledge: knowledge_integration (h/c), knowledge_version (h/c), graph_query, graph_reasoning, knowledge.c, CMakeLists.txt
- Reasoning: reasoning.c, reasoning_internal, planning, CMakeLists.txt
- Robot: computer_operation (h/c), robot_agent (h/c), hardware_resource_manager, robot.c, ros_gazebo_bridge, ros_protocol, simulator, CMakeLists.txt
- Learning: imitation_learning, meta_learning, multi_agent, online_learning, reinforcement_learning
- GPU: gpu_benchmark, gpu_cpu, gpu_tpu
- Multimodal: slam.h, data_collection_pipeline, image_recognition_deep, liquid_vision, ocr, point_cloud, speech_recognition, text
- AGI: agi.c
- Cognition: metacognition, self_cognition (h/c)
- Math: matrix_ops.h
- Programming: self_programming.h
- Memory: memory.c
- Distributed: pbft
- Multisystem: swarm_intelligence
- Evolution: evolution_engine
- Backend: backend.c, CMakeLists.txt
- Training: training.c, CMakeLists.txt
- CMakeLists.txt, main.c, utils/logging

+++ src/distributed/pbft.c
@@ -1012,7 +1012,7 @@ static int pbft_send_view_change(PbftSystem* system, uint32_t new_view_number) {
     vc.prepared_count = 0;
     /* 收集当前视图下已prepared但未committed的证据 */
     for (int i = 0; i < system->log_count && vc.prepared_count < PBFT_MAX_PENDING_BATCH; i++) {
-        if (system->log_entries[i].prepare_count >= (uint32_t)(2 * system->config.max_fault)) {
+        if (system->log_entries[i].prepare_count >= (int)(2U * (uint32_t)system->config.max_fault)) {
             vc.prepared_evidence[vc.prepared_count].seq_number = system->log_entries[i].sequence_number;
             memcpy(vc.prepared_evidence[vc.prepared_count].digest,
                    system->log_entries[i].request_digest, sizeof(vc.prepared_evidence[0].digest));
